Restructure files and components and remove unused code
