--- src/App.tsx
@@ -1,7 +1,0 @@
-import EvalBar from './EvalBar'
-
-function App() {
-  return <EvalBar />
-}
-
-export default App
