+++ src/eval-bar/eval-bar.tsx
@@ -1,6 +1,6 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
+import { EvalBarDisplay } from '@/eval-bar/eval-bar-display'
+import { Prediction } from '@/eval-bar/types'
 import { useEffect, useRef, useState } from 'react'
-import { Prediction } from './types'
 
 export default function EvalBar() {
   const [evalValue, setEvalValue] = useState(50)
@@ -109,42 +109,5 @@ export default function EvalBar() {
     return () => clearInterval(interval.current)
   }, [])
 
-  return (
-    <div
-      style={{
-        border: '0px solid red',
-        display: 'flex',
-        backgroundColor: 'rgba(94, 130, 191, 0.5)',
-        boxSizing: 'border-box',
-        height: '100vh',
-        flexDirection: 'column',
-        justifyContent: 'stretch',
-      }}
-    >
-      <div
-        style={{
-          flexGrow: 100 - evalValue,
-          backgroundColor: '#db00b3',
-          color: '#db00b3',
-          fontWeight: 'bold',
-          transition: 'all 1s',
-        }}
-      >
-        Doubters
-      </div>
-      <div
-        style={{
-          flexGrow: evalValue,
-          backgroundColor: '#1e69ff',
-          color: '#1e69ff',
-          fontWeight: 'bold',
-          alignContent: 'end',
-          transition: 'all 1s',
-        }}
-      >
-        Believers
-      </div>
-      {error && <p style={{ color: 'red' }}>Error: {error}</p>}
-    </div>
-  )
+  return <EvalBarDisplay evalValue={evalValue} error={error} />
 }
